refactor: list notebooks
